requirements.txt
numpy==1.17.5
argparse
pathlib
pillow
